+++ stencil.config.dist.ts
@@ -1,5 +1,6 @@
 import { Config } from '@stencil/core';
 import { sass } from '@stencil/sass';
+import nodePolyfills from 'rollup-plugin-node-polyfills';
 
 export const config: Config = {
     namespace: 'lime-elements',
@@ -13,6 +14,9 @@ export const config: Config = {
         },
     ],
     plugins: [sass()],
+    rollupPlugins: {
+        after: [nodePolyfills()],
+    },
     tsconfig: './tsconfig.dist.json',
     globalStyle: 'src/global/core-styles.scss',
 };